+++ CHANGES.rst
@@ -24,6 +24,7 @@ Changelog
 - When advice is added/modified/removed, clean the `Products.PloneMeeting.MeetingItem.modified` cachekey volatile to clear cache for portlet_todo
 - Adapted the way late items work: now an item is late for the selected preferred meeting and for every following meetings.  This way an item that was late for a meeting may also
   be presented as late item for next meeting instead only being presentable to next non frozen meeting
+- Moved `MeetingItemWorkflowConditions._groupIsNotEmpty` to `ToolPloneMeeting.group_is_not_empty` so it is easier to use everywhere
 - Added new field MeetingItem.meetingManagersNotes only viewable/editable by MeetingManagers
 
 4.1.18 (2020-02-21)